requirements.txt
PyYAML==6.0.2
Jinja2==3.1.4
rich==13.9.2
tqdm==4.66.5
yara-python==4.5.1
